Moved /controls/ out of /component/. Added new Checkbox.js control.
