+++ src/components/column/EditColumn.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect, useContext } from "react";
 import { AppContext } from "../../AppContextProvider";
 import { columnsClient } from "../../api/columnsClient";
 import { handleError } from "../../util/handleError";
-import { Input } from "../controls/inputs/Input";
-import { PrimaryButton } from "../controls/buttons/PrimaryButton";
+import { Input } from "../../controls/inputs/Input";
+import { PrimaryButton } from "../../controls/buttons/PrimaryButton";
 import KeyboardBackspaceIcon from '@mui/icons-material/KeyboardBackspace';
 
 export const EditColumn = ({ setFormError, setError, closeSideModal, handleRerender, columnId }) => {
