+++ src/components/task/ViewTask.js
@@ -4,7 +4,7 @@ import { tasksClient } from "../../api/tasksClient";
 import { handleError } from "../../util/handleError";
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
-import { FlatButton } from "../controls/buttons/FlatButton";
+import { FlatButton } from "../../controls/buttons/FlatButton";
 import "./ViewTask.css"
 
 export const ViewTask = ({ taskId, openEditTaskModal, setError, handleRerender }) => {
