+++ src/controls/buttons/DeleteButton.js
