+++ src/controls/icons/AddIcon.js
