+++ src/controls/icons/MoreIcon.css
