+++ src/controls/icons/MoreIcon.js
