+++ src/controls/icons/ProjectBLogo.js
